requirements.txt
openpyxl
pandas
playwright
google-cloud-storage
google-cloud-bigquery
gspread
oauth2client  # if using Google Sheets API
